LayoutTests/fast/events/popup-blocking-timers.html

<head>
    <script src="../js/resources/js-test-pre.js"></script>
    <script>
        var newWindow;
        var intervalId;
        var firstIntervalExecution = true;
        var clickNumber = 0;
        
        if (window.testRunner) {
            testRunner.dumpAsText();
            testRunner.setCanOpenWindows();
            testRunner.waitUntilDone();
            testRunner.setPopupBlockingEnabled(true);
        }
    
        function clickHandler() {
            clickNumber++;
            if (clickNumber == 1) {
                newWindow = window.open("about:blank");
                self.focus();
                debug("Test calling window.open() directly. A popup should be allowed.");
                shouldBeNonNull("newWindow");
            } else if (clickNumber == 2) {
                setTimeout(function() {
                    newWindow = window.open("about:blank");
                    self.focus();
                    debug("Test calling window.open() with a 0 ms delay. A popup should be allowed.")
                    shouldBeNonNull("newWindow");
                }, 0);
            } else if (clickNumber == 3) {
                intervalId = setInterval(function() {
                    debug("Test calling window.open() in a 100 ms interval. A popup should only be allowed on the first execution of the interval.");
                    newWindow = window.open("about:blank");
                    self.focus();
                    if (firstIntervalExecution) {
                        shouldBeNonNull("newWindow");
                        firstIntervalExecution = false;
                    } else {
                        shouldBeUndefined("newWindow");
                        clearInterval(intervalId);
                    }
                }, 100);
            } else if (clickNumber == 4) {
                setTimeout(function() {
                    setTimeout(function() {
                        newWindow = window.open("about:blank");
                        self.focus();
                        debug("Test calling window.open() in a nested call to setTimeout(). A popup should not be allowed.")
                        shouldBeUndefined("newWindow");
                    }, 0);
                }, 300);
            } else if (clickNumber == 5) {
                setTimeout(function() {
                    newWindow = window.open("about:blank");
                    self.focus();
                    debug("Test calling window.open() with a 1000 ms delay. A popup should be allowed.")
                    shouldBeNonNull("newWindow");
                }, 1000);
            } else if (clickNumber == 6) {
                setTimeout(function() {
                    newWindow = window.open("about:blank");
                    self.focus();
                    debug("Test calling window.open() with a 1001 ms delay. A popup should not be allowed.")
                    shouldBeUndefined("newWindow");

                    if (window.testRunner)
                        testRunner.notifyDone();
                }, 1001);
            }
            if (window.eventSender)
                eventSender.leapForward(1001);
        }
        
        function clickButton() {
            var button = document.getElementById("test");
            var buttonX = button.offsetLeft + button.offsetWidth / 2;
            var buttonY = button.offsetTop + button.offsetHeight / 2;
            if (window.eventSender) {
                for (var i = 0; i < 6; i++) { 
                    eventSender.mouseMoveTo(buttonX, buttonY);
                    eventSender.mouseDown();
                    eventSender.mouseUp();
                }
            }
        }        
    </script>
</head>
<body onload="clickButton()">
    <button id="test" onclick="clickHandler()">Click Here (6 times)</button>
    <div id="console"></div>
</body>
